Add git-replacepr
This script can be used to entirely replace the contents of the
underlying branch with a new commit. This is useful if you've been
iterating on a PR, but then either want to squash it, or you for some
reason changed things on the main branch without mirroring those on the
remote branch.

(cherry picked from commit 06b96c8f11a6a1f9f882599ac87d4b7a8a1ebcfe)